Update segmentAnalysis.py
This has the desired effect, but takes a very long time. This could be sped up dramatically if the number of points was narrowed down. Many of the points are not needed. Maybe filter by the coordinates of the points? If not in a bounding box of the track, move on? Getting pretty close to just splitting up the analysis completely based on trackId moving in this direction, which would be helpful for the points version as well as this version.